+++ src/lib/flarmgatas/ace/flarm2024.hpp
@@ -48,7 +48,7 @@ public:
     Flarm2024(etl::imessage_bus &bus, const Configuration &config) : BaseModule(bus, NAME)
     {
         auto di = config.valueByPath(DEFAULT_IGNORE_DISTANCE, "Flarm", "distanceIgnore");
-        distanceIgnore = etl::max(0, etl::min(di, MAX_IGNORE_DISTANCE));
+        distanceIgnore = static_cast<uint32_t>(etl::max(0, etl::min(di, MAX_IGNORE_DISTANCE)));
         gaTasConfiguration = config.gaTasConfig();
     }
